Fix bug with date range in mp-calendar
The bug happens in the following case:
1. Click a date in calendar to pick start of date range (e.g. Feb 1, 2017)
2. Click a date in calendar to pick end of date range (e.g. Feb 9, 2017)
3. Update the selected-start-date to a start date after the current end date
(e.g. Feb 13, 2017)

Expected:
- Only Feb 13, 2017 should be selected (it's both the new start and end
date)

Actual:
- Now, the calendar will have both Feb 9 and Feb 13 selected, without
any days in between filled in.

The reason this happens is because clicking a date in the calendar
implicitly calls setDate(...) on the clicked date. This date cell is
given the class `is-selected` and gets the selected UI state. However,
setting the attribute `selected-from-date` or `selected-to-date` does
not call `setDate(...)`. One fix is to call `setDate(...)` with the updated
`selected-from-date` or `selected-to-date` whenever one of them is
updated. However, it's possible that the caller updates both
`seleted-from-date` and `selected-to-date` together, in which case it's
non-deterministic which becomes the selected date. Instead, I've opted to
not having a selected date when we're working with a range by calling
`setDate(null)`.